reformat the output
The test output is purposely rather verbose, it prints the commands
being executed, their output at times and the checks for it.

This changes the output to make use of more colors and groups blocks
of actions, typically started with a short description (white) and
ends with the checks (red or green). program output is now gray, and
commands blue. Warnings are now yellow.